# Optimize cupy.asfortranarray