--- a/tests/cupy_tests/manipulation_tests/test_kind.py
+++ b/tests/cupy_tests/manipulation_tests/test_kind.py
@@ -40,3 +40,23 @@ class TestKind(unittest.TestCase):
             self.assertTrue(ret.flags.f_contiguous)
             return ret.strides
         self.assertEqual(func(numpy), func(cupy))
+
+    @testing.for_all_dtypes()
+    def test_asfortranarray4(self, dtype):
+        def func(xp):
+            x = xp.zeros((2, 3), dtype)
+            x = xp.transpose(x, (1, 0))
+            ret = xp.asfortranarray(x)
+            self.assertTrue(ret.flags.f_contiguous)
+            return ret.strides
+        self.assertEqual(func(numpy), func(cupy))
+
+    @testing.for_all_dtypes()
+    def test_asfortranarray5(self, dtype):
+        def func(xp):
+            x = testing.shaped_arange((2, 3), xp, dtype)
+            ret = xp.asfortranarray(x)
+            self.assertTrue(x.flags.c_contiguous)
+            self.assertTrue(ret.flags.f_contiguous)
+            return ret.strides
+        self.assertEqual(func(numpy), func(cupy))