@rezoo 2017-01-10
@okuta If you do not find any problems, I would like to merge this PR.